изменил структуру приложения, прокинул массив posts через пропсы
