+++ src/components/Dialogs/Message/Message.tsx
@@ -1,14 +1,9 @@
 import React from 'react'
 import s from '../Dialogs.module.css';
 
-
-
 type MessageType = {
     message: string
 }
-
 export const Message = (props: MessageType) =>{
     return  <div className={s.message}>{props.message}</div>
 }
-
-
